+++ vk-video/src/codec.rs
@@ -38,7 +38,10 @@ pub(crate) trait EncodeCodec: Codec {
     fn vk_parameters<'a>(parameters: &'a Self::OwnedParameters) -> Self::VkParameters<'a>;
 
     type BitstreamUnitData;
-    fn bitstream_unit_data(is_idr: bool) -> Self::BitstreamUnitData;
+    fn bitstream_unit_data(
+        codec_capabilities: &Self::CodecSpecificEncodeCapabilities<'_>,
+        is_idr: bool,
+    ) -> Self::BitstreamUnitData;
     type BitstreamUnitInfo<'a>;
     fn bitstream_unit_info<'a>(
         data: &'a Self::BitstreamUnitData,
@@ -61,6 +64,7 @@ pub(crate) trait EncodeCodec: Codec {
     type PictureInfoData;
     fn picture_info_data(
         counters: &Self::EncodingCounters,
+        codec_capabilities: &Self::CodecSpecificEncodeCapabilities<'_>,
         is_idr: bool,
         ref_lists: &Self::ReferenceListInfo,
     ) -> Self::PictureInfoData;