+++ vk-video/src/vulkan_video.rs
@@ -74,7 +74,7 @@ pub mod parameters {
     }
 
     impl H265Profile {
-        pub(crate) fn to_profile_idc(self) -> ash::vk::native::StdVideoH264ProfileIdc {
+        pub(crate) fn to_profile_idc(self) -> ash::vk::native::StdVideoH265ProfileIdc {
             match self {
                 H265Profile::Main => {
                     ash::vk::native::StdVideoH265ProfileIdc_STD_VIDEO_H265_PROFILE_IDC_MAIN